+++ src/app/models/product-cart-info.model.ts
@@ -1,9 +1,10 @@
-import { Product } from '.';
+import { Product } from './product.model';
 
 export class ProductCartInfo {
     public id: number;
     public name: string;
     public price: number;
+    public total: number;
 
     constructor(
         product: Product,
@@ -12,6 +13,11 @@ export class ProductCartInfo {
         this.id = product.id;
         this.name = product.name;
         this.price = product.price;
-        this.count = (count !== undefined) ? count : 1;
+        this.total = product.count;
+        this.count = (count !== undefined) ? count : 0;
+    }
+
+    public get notAvailable(): boolean {
+        return this.count >= this.total;
     }
 }