+++ src/app/modules/index.ts
@@ -1,0 +1,2 @@
+export * from './product/product.module';
+export * from './cart/cart.module';